backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# Validation & Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database drivers
aiosqlite==0.19.0
asyncpg==0.29.0
greenlet==3.0.3

# Image processing (for receipt OCR placeholder)
Pillow==10.2.0

# Email (for verification - optional)
email-validator==2.1.0
